--- Sedes_y_centros_de_acopio/carrito.json
@@ -1,17 +1,0 @@
-[
-    {
-        "nombre": "Material 2",
-        "cantidad": "300 Kilogramo",
-        "tec_colones": 300.0
-    },
-    {
-        "nombre": "shasohb",
-        "cantidad": "100 Unidad",
-        "tec_colones": 3300.0
-    },
-    {
-        "nombre": "weqqew",
-        "cantidad": "60 Kilogramo",
-        "tec_colones": 120.0
-    }
-]
